requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
importlib-metadata==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
